requirements_full.txt
gitpython
numpy
sympy

# used for query available memory
psutil

# used for print table
tabulate

pytest
tqdm

# used to annotate the scope of events in host process, which can be visualized in Nsight System.
nvtx

# for onnx frontend
onnx
onnxruntime-gpu

# for tensor rt baseline
# nvidia-pyindex
--extra-index-url https://pypi.ngc.nvidia.com
tensorrt

# for tvm tuning
xgboost
tornado
cloudpickle
